# src/main/java/com/wtf/cauth/data/dto/request/app/RefreshAppSecretReqDto.java
package com.wtf.cauth.data.dto.request.app;

import com.fasterxml.jackson.databind.PropertyNamingStrategy.SnakeCaseStrategy;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@JsonNaming(SnakeCaseStrategy.class)
public class RefreshAppSecretReqDto {
    private String name;
    private String curSecret;
}