# src/main/java/com/wtf/cauth/data/dto/request/user/AuthTokenReqDto.java
package com.wtf.cauth.data.dto.request.user;

import com.fasterxml.jackson.databind.PropertyNamingStrategy.SnakeCaseStrategy;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@JsonNaming(SnakeCaseStrategy.class)
public class AuthTokenReqDto {
    private String authToken;
}