# src/main/java/com/wtf/cauth/data/dto/response/user/AuthTokenResDto.java
package com.wtf.cauth.data.dto.response.user;

import com.fasterxml.jackson.databind.PropertyNamingStrategy.SnakeCaseStrategy;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@JsonNaming(SnakeCaseStrategy.class)
public class AuthTokenResDto {
    private boolean valid;
}