# src/main/java/com/wtf/cauth/data/dto/response/user/UserLoginResDto.java
package com.wtf.cauth.data.dto.response.user;

import com.fasterxml.jackson.databind.PropertyNamingStrategy.SnakeCaseStrategy;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@JsonNaming(SnakeCaseStrategy.class)
public class UserLoginResDto {
    private String authToken;
    private long expiresOn;
    private String id;
    private String role;
    private String appName;
}